nfdc: add fib list
