+++ tools/nfdc.go
@@ -1,14 +1,10 @@
 package tools
 
 import (
-	"fmt"
-	"time"
-
 	enc "github.com/named-data/ndnd/std/encoding"
 	"github.com/named-data/ndnd/std/engine"
 	"github.com/named-data/ndnd/std/log"
 	"github.com/named-data/ndnd/std/ndn"
-	mgmt "github.com/named-data/ndnd/std/ndn/mgmt_2022"
 	"github.com/named-data/ndnd/std/utils"
 )
 
@@ -30,8 +26,12 @@ func GetNfdcCmdTree() utils.CmdTree {
 			Fun:  nfdc.ExecRouteList,
 		}, {
 			Name: "route add",
-			Help: "Control RIB and FIB entries",
+			Help: "Add a route to the RIB",
 			Fun:  execCmd("rib", "add"),
+		}, {
+			Name: "fib list",
+			Help: "Print FIB entries",
+			Fun:  nfdc.ExecFibList,
 		}},
 	}
 }
@@ -60,90 +60,4 @@ func (n *Nfdc) GetPrefix() enc.Name {
 		enc.NewStringComponent(enc.TypeGenericNameComponent, "localhost"),
 		enc.NewStringComponent(enc.TypeGenericNameComponent, "nfd"),
 	}
-}
-
-func (n *Nfdc) ExecCmd(mod string, cmd string, args []string) {
-	n.Start()
-	defer n.Stop()
-
-	argmap := map[string]any{}
-	ctrlArgs, err := mgmt.DictToControlArgs(argmap)
-	if err != nil {
-		log.Fatalf("Invalid control args: %+v", err)
-		return
-	}
-
-	resp, err := n.engine.ExecMgmtCmd(mod, cmd, ctrlArgs)
-	if err != nil {
-		log.Fatalf("Error executing command: %+v", err)
-		return
-	}
-
-	log.Infof("Response: %+v", resp)
-}
-
-func (n *Nfdc) ExecRouteList(args []string) {
-	suffix := enc.Name{
-		enc.NewStringComponent(enc.TypeGenericNameComponent, "rib"),
-		enc.NewStringComponent(enc.TypeGenericNameComponent, "list"),
-	}
-
-	data, err := n.fetchStatusDataset(suffix)
-	if err != nil {
-		log.Fatalf("Error fetching status dataset: %+v", err)
-		return
-	}
-
-	status, err := mgmt.ParseRibStatus(enc.NewWireReader(data), true)
-	if err != nil {
-		log.Fatalf("Error parsing RIB status: %+v", err)
-		return
-	}
-
-	for _, entry := range status.Entries {
-		for _, route := range entry.Routes {
-			expiry := "never"
-			if route.ExpirationPeriod != nil {
-				expiry = (time.Duration(*route.ExpirationPeriod) * time.Millisecond).String()
-			}
-
-			// TODO: convert origin, flags to string
-			fmt.Printf("prefix=%s nexthop=%d origin=%d cost=%d flags=%d expires=%s\n",
-				entry.Name, route.FaceId, route.Origin, route.Cost, route.Flags, expiry)
-		}
-	}
-}
-
-func (n *Nfdc) fetchStatusDataset(suffix enc.Name) (enc.Wire, error) {
-	n.Start()
-	defer n.Stop()
-
-	// TODO: segmented fetch once supported by fw/mgmt
-	name := append(n.GetPrefix(), suffix...)
-	config := &ndn.InterestConfig{
-		MustBeFresh: true,
-		CanBePrefix: true,
-		Lifetime:    utils.IdPtr(time.Second),
-		Nonce:       utils.ConvertNonce(n.engine.Timer().Nonce()),
-	}
-	interest, err := n.engine.Spec().MakeInterest(name, config, nil, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	ch := make(chan ndn.ExpressCallbackArgs)
-	err = n.engine.Express(interest, func(args ndn.ExpressCallbackArgs) {
-		ch <- args
-		close(ch)
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	res := <-ch
-	if res.Result != ndn.InterestResultData {
-		return nil, fmt.Errorf("interest failed: %d", res.Result)
-	}
-
-	return res.Data.Content(), nil
 }
